requirements.txt
numpy>=2.2.4
openpyxl>=3.1.5
pandas>=2.2.3
pydantic>=2.11.2
pydantic-settings>=2.8.1
pydantic_core>=2.33.1
PyQt5>=5.15.11
PyYAML>=6.0.2
tinydb>=4.8.2
tinydb-serialization>=2.2.0
pyinstaller
